+++ tests/testthat/test-00.updatePlantumlJar.R
@@ -1,4 +1,12 @@
-context("check update process")
+context("00 check update process")
+
+test_that(
+  "Automatic download when calling plantuml_run()",
+  expect_error(
+    object = plantuml_run(stdout = TRUE),
+    regexp = NA
+  )
+)
 
 test_that(
   "updatePlantumlJar(beta = TRUE)",